requirements.txt
beautifulsoup4>=4.12.3
dateparser>=1.2.0
flask>=3.0.3
google-api-python-client>=2.126.0
google-auth-httplib2>=0.2.0
google_auth_oauthlib>=1.2.0
gTTS>=2.5.1
matplotlib>=3.8.4
msal>=1.28.0
obd>=0.7.2
openai>=1.21.2
pandas>=2.2.2
pygame>=2.5.2
pyserial>=3.5
PyAudio>=0.2.14
python-dotenv>=1.0.1
pyttsx4>=3.0.15
pytz>=2024.1
requests>=2.31.0
rich>=13.7.1
spacy>=3.7.4
SpeechRecognition>=3.10.3
twilio>=9.0.4